+++ src/Silverback.Integration/Messaging/Messages/InboundEnvelope.cs
@@ -56,4 +56,6 @@ internal record InboundEnvelope : RawInboundEnvelope, IInboundEnvelope
     public bool AutoUnwrap => true;
 
     public object? Message { get; set; }
+
+    public bool IsTombstone => Message is null or ITombstone;
 }